+++ ambry-api/src/main/java/com/github/ambry/quota/EnforcementRecommendation.java
@@ -19,36 +19,73 @@ package com.github.ambry.quota;
  * request or not, along with usage information like usage percentage, name of the enforcer that made this recommendation,
  * the recommended http status (indicating whether or not throttled request should be retried) and cost to serve the request.
  */
-public interface EnforcementRecommendation {
+public class EnforcementRecommendation {
+  private final boolean shouldThrottle;
+  private final float usagePercentage;
+  private final QuotaName quotaName;
+  private final int recommendedHttpStatus;
+  private final double requestCost;
+  private final long retryAfterMs;
+
+  /**
+   * Constructor for {@link EnforcementRecommendation}.
+   * @param shouldThrottle boolean flag indicating throttling recommendation.
+   * @param usagePercentage percentage of resource usage.
+   * @param quotaName name of the enforcement that made the recommendation.
+   * @param recommendedHttpStatus recommended http status.
+   * @param requestCost cost of the request.
+   * @param retryAfterMs time after which request can be retried.
+   */
+  public EnforcementRecommendation(boolean shouldThrottle, float usagePercentage, QuotaName quotaName,
+      int recommendedHttpStatus, float requestCost, long retryAfterMs) {
+    this.shouldThrottle = shouldThrottle;
+    this.usagePercentage = usagePercentage;
+    this.quotaName = quotaName;
+    this.recommendedHttpStatus = recommendedHttpStatus;
+    this.requestCost = requestCost;
+    this.retryAfterMs = retryAfterMs;
+  }
 
   /**
    * @return true if recommendation is to throttle. false otherwise.
    */
-  boolean shouldThrottle();
+  public boolean shouldThrottle() {
+    return shouldThrottle;
+  }
 
   /**
    * @return estimation of percentage of quota in use when the recommendation was made.
    */
-  float getQuotaUsagePercentage();
+  public float getQuotaUsagePercentage() {
+    return usagePercentage;
+  }
 
   /**
    * @return name of the quota for which recommendation is made.
    */
-  QuotaName getQuotaName();
+  public QuotaName getQuotaName() {
+    return quotaName;
+  }
 
   /**
    * @return http status recommended by enforcer.
    */
-  int getRecommendedHttpStatus();
+  public int getRecommendedHttpStatus() {
+    return recommendedHttpStatus;
+  }
 
   /**
    * @return the cost of serving the request.
    */
-  double getRequestCost();
+  public double getRequestCost() {
+    return requestCost;
+  }
 
   /**
    * @return the time interval in milliseconds after the request can be retried.
    * If request is not throttled then returns -1.
    */
-  long getRetryAfterMs();
+  public long getRetryAfterMs() {
+    return retryAfterMs;
+  }
 }
