+++ ambry-api/src/main/java/com/github/ambry/quota/QuotaMode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 LinkedIn Corp. All rights reserved.
+ * Copyright 2020 LinkedIn Corp. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
