+++ ambry-api/src/main/java/com/github/ambry/quota/QuotaResource.java
@@ -48,6 +48,6 @@ public class QuotaResource {
    * Type of Ambry resource for which quota can be applied.
    */
   enum QuotaResourceType {
-    AMBRY_ACCOUNT, AMBRY_CONTAINER, AMBRY_HOST
+    AMBRY_ACCOUNT, AMBRY_CONTAINER
   }
 }
