+++ ambry-quota/src/main/java/com/github/ambry/quota/MaxThrottlePolicy.java
@@ -18,6 +18,11 @@ import java.util.List;
 import java.util.Map;
 
 
+/**
+ * An implementation of {@link ThrottlePolicy} that creates a {@link ThrottlingRecommendation} to throttle if any one of
+ * {@link EnforcementRecommendation} recommendation is to throttle, and takes the max of retry after time interval. Also
+ * groups the quota usage and request cost for all the quotas.
+ */
 public class MaxThrottlePolicy implements ThrottlePolicy {
   @Override
   public ThrottlingRecommendation recommend(List<EnforcementRecommendation> enforcementRecommendations) {
@@ -33,7 +38,7 @@ public class MaxThrottlePolicy implements ThrottlePolicy {
       requestCost.put(recommendation.getQuotaName(), recommendation.getRequestCost());
       retryAfterMs = Math.max(recommendation.getRetryAfterMs(), retryAfterMs);
     }
-    return new AmbryThrottlingRecommendation(shouldThrottle, quotaUsagePercentage, recommendedHttpStatus, requestCost,
+    return new ThrottlingRecommendation(shouldThrottle, quotaUsagePercentage, recommendedHttpStatus, requestCost,
         retryAfterMs);
   }
 }
