+++ ambry-test-utils/src/main/java/com/github/ambry/quota/RejectRequestQuotaEnforcer.java
@@ -40,13 +40,13 @@ public class RejectRequestQuotaEnforcer implements RequestQuotaEnforcer {
 
   @Override
   public EnforcementRecommendation chargeAndRecommend(RestRequest restRequest, BlobInfo blobInfo) {
-    return new AmbryEnforcementRecommendation(SHOULD_THROTTLE, DUMMY_REJECTABLE_USAGE_PERCENTAGE,
+    return new EnforcementRecommendation(SHOULD_THROTTLE, DUMMY_REJECTABLE_USAGE_PERCENTAGE,
         QuotaName.READ_CAPACITY_UNIT, REJECT_HTTP_STATUS, 1, -1);
   }
 
   @Override
   public EnforcementRecommendation recommend(RestRequest restRequest) {
-    return new AmbryEnforcementRecommendation(SHOULD_THROTTLE, DUMMY_REJECTABLE_USAGE_PERCENTAGE,
+    return new EnforcementRecommendation(SHOULD_THROTTLE, DUMMY_REJECTABLE_USAGE_PERCENTAGE,
         QuotaName.READ_CAPACITY_UNIT, REJECT_HTTP_STATUS, 1, -1);
   }
